Manipulate nested Times/NonCommutativeMultiply/Power more globally
The "local" approach taken previously led to infinite loops: for
example if a**b==a*b and b**c==b*c and c**d==c*d then
(a*b)**(c*d) was expanded to a**(b*c)**d because b**c==b*c, then
this was expanded to the original (a*b)**(c*d) because a**b==a*b.
Now terms are normalized to a fixed form to avoid such loops.